+++ public/controllers/management/edition.js
@@ -94,7 +94,7 @@ export class EditionController {
         this.$scope.fetchedXML = null;
         this.errorHandler.handle(error, 'Fetch file error');
       }
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
     };
 
     this.$scope.restartNode = async selectedNode => {
@@ -165,7 +165,7 @@ export class EditionController {
 
     this.$scope.xmlIsValid = valid => {
       this.$scope.xmlHasErrors = valid;
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
     };
 
     this.$scope.changeNode = () => {
@@ -202,6 +202,6 @@ export class EditionController {
     }
 
     this.$scope.editConf();
-    if (!this.$scope.$$phase) this.$scope.$digest();
+    this.$scope.$applyAsync();
   }
 }