+++ public/controllers/settings/settings.js
@@ -146,7 +146,7 @@ export class SettingsController {
       for (const key in this.showEditForm) this.showEditForm[key] = false;
       this.$scope.$emit('updateAPI', {});
       this.errorHandler.info('The API was removed successfully', 'Settings');
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
       return;
     } catch (error) {
       this.errorHandler.handle('Could not remove the API', 'Settings');
@@ -218,7 +218,7 @@ export class SettingsController {
 
     this.extensions = this.appState.getExtensions(JSON.parse(currentApi).id);
 
-    if (!this.$scope.$$phase) this.$scope.$digest();
+    this.$scope.$applyAsync();
     return;
   }
 
@@ -250,7 +250,7 @@ export class SettingsController {
         this.currentDefault = JSON.parse(currentApi).id;
       }
 
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
       this.getCurrentAPIIndex();
       if (!this.currentApiEntryIndex && this.currentApiEntryIndex !== 0) return;
 
@@ -266,7 +266,7 @@ export class SettingsController {
 
       this.extensions = this.appState.getExtensions(JSON.parse(currentApi).id);
 
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
       return;
     } catch (error) {
       this.errorHandler.handle('Error getting API entries', 'Settings');
@@ -325,7 +325,7 @@ export class SettingsController {
     this.showEditForm[entry._id] = !this.showEditForm[entry._id];
     this.isEditing = this.showEditForm[entry._id];
     this.addManagerContainer = false;
-    if (!this.$scope.$$phase) this.$scope.$digest();
+    this.$scope.$applyAsync();
   }
 
   // Save settings function
@@ -451,7 +451,7 @@ export class SettingsController {
       this.printError(error);
     }
     this.savingApi = false;
-    if (!this.$scope.$$phase) this.$scope.$digest();
+    this.$scope.$applyAsync();
     return;
   }
 
@@ -521,7 +521,7 @@ export class SettingsController {
       this.printError(error, true);
     }
     this.savingApi = false;
-    if (!this.$scope.$$phase) this.$scope.$digest();
+    this.$scope.$applyAsync();
     return;
   }
 
@@ -561,7 +561,7 @@ export class SettingsController {
       this.apiIsDown = false;
       this.errorHandler.info('Connection success', 'Settings');
 
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
       return;
     } catch (error) {
       if (!this.wzMisc.getApiIsDown()) this.printError(error);
@@ -582,7 +582,7 @@ export class SettingsController {
       this.apiEntries[
         this.currentApiEntryIndex
       ]._source.extensions = currentExtensions;
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
     } catch (error) {
       this.errorHandler.handle(error, 'Settings');
     }
@@ -606,7 +606,7 @@ export class SettingsController {
         'Settings'
       );
       this.selectedIndexPattern = newIndexPattern;
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
       return;
     } catch (error) {
       this.errorHandler.handle(
@@ -637,7 +637,7 @@ export class SettingsController {
       const logs = await this.genericReq.request('GET', '/utils/logs', {});
       this.logs = logs.data.lastLogs.map(item => JSON.parse(item));
       this.loadingLogs = false;
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
     } catch (error) {
       this.logs = [
         {
@@ -696,7 +696,7 @@ export class SettingsController {
       if (this.currentApiEntryIndex || this.currentApiEntryIndex === 0) {
         await this.checkManager(this.currentApiEntryIndex, true);
       }
-      if (!this.$scope.$$phase) this.$scope.$digest();
+      this.$scope.$applyAsync();
       return;
     } catch (error) {
       this.errorHandler.handle(