+++ public/directives/wz-table/lib/sort.js
@@ -24,6 +24,6 @@ export async function sort(field, $scope, instance, fetch, errorHandler) {
     $scope.error = errorHandler.handle(error.message || error, 0, 0, 1);
     errorHandler.handle(error.message || error);
   }
-  if (!$scope.$$phase) $scope.$digest();
+  $scope.$applyAsync();
   return;
 }