+++ public/kibana-integrations/kibana-visualization.js
@@ -192,8 +192,8 @@ app.directive('kbnVis', function() {
       };
 
       // Listen for changes
-      const updateVisWatcher = $rootScope.$on('updateVis', () => {
-        if (!$rootScope.$$phase) $rootScope.$digest();
+      const updateVisWatcher = $scope.$on('updateVis', () => {
+        $scope.$applyAsync();
         const rawVis = rawVisualizations.getList();
         if (Array.isArray(rawVis) && rawVis.length) {
           myRender(rawVis);
@@ -230,7 +230,7 @@ app.directive('kbnVis', function() {
           $rootScope.rendered = false;
         }
         // Forcing a digest cycle
-        if (!$rootScope.$$phase) $rootScope.$digest();
+        $rootScope.$applyAsync();
       };
 
       // Initializing the visualization